Refresh submission message timeout on button press
Change the way that the timeout is handled for the "Settings have been
sent to GoPiGo" message so that the timeout is refreshed each time the
submit button is pressed.

This prevents issues with the submit message disappearing add odd times
due to a previous click tiemout kicking in.

Also remove the submit modal.